webselenium/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
idna==2.8
isort==4.3.21
Jinja2==2.9.5
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
PyYAML==5.1.1
requests==2.22.0
rope==0.14.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
selenium==3.141.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
BeautifulReport==0.1.2
